+++ src/net_prof/__init__.py
@@ -1,7 +1,7 @@
 # src/net_prof/__init__.py
 
 from .engine import summarize, collect, dump, dump_html, collect
-from .visualize import bar_chart. heat_map, generate_iface_barchart
+from .visualize import bar_chart, heat_map, generate_iface_barchart
 
 __all__ = [
     "summarize",